separated client and server rendering
