+++ app/react/server.js
@@ -1,4 +1,4 @@
-import Router from './Router';
+import Router from './ServerRouter.js';
 
 export default app => {
   app.get(/^\/(?!api(\/|$)).*$/, Router);
